[test] Parameterize tests over [serverless, pod]
I've created a centralized conftest.py across all tests (e2e and system) with parametrized fixtures - index_params and namespace. Now all tests can simply depend on these fixtures to be parameterized
