+++ tests/unit/chat_engine/test_chat_engine.py
@@ -13,7 +13,6 @@ from canopy.llm import BaseLLM
 from canopy.models.api_models import ChatResponse, _Choice, TokenCounts
 from canopy.models.data_models import MessageBase, Query, Context, Role
 from .. import random_words
-from ...util import TEST_NAMESPACE
 
 MOCK_SYSTEM_PROMPT = "This is my mock prompt"
 MAX_PROMPT_TOKENS = 100
@@ -101,9 +100,6 @@ class TestChatEngine:
         }
         return messages, expected
 
-    @pytest.mark.parametrize("namespace", [
-        None, TEST_NAMESPACE
-    ])
     def test_chat(self, namespace, history_length=5, snippet_length=10):
 
         chat_engine = self._init_chat_engine()
@@ -137,9 +133,6 @@ class TestChatEngine:
 
         assert response == expected['response']
 
-    @pytest.mark.parametrize("namespace", [
-        None, TEST_NAMESPACE
-    ])
     def test_chat_engine_params(self,
                                 namespace,
                                 system_prompt_length=10,
